Revert "Спринт 14."
