+++ .idea/kotlinc.xml
@@ -1,12 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
-  <component name="Kotlin2JvmCompilerArguments">
-    <option name="jvmTarget" value="1.8" />
-  </component>
-  <component name="KotlinCommonCompilerArguments">
-    <option name="apiVersion" value="1.9" />
-    <option name="languageVersion" value="1.9" />
-  </component>
   <component name="KotlinJpsPluginSettings">
     <option name="version" value="1.9.0" />
   </component>
