+++ app/src/main/java/com/practicum/playlistmaker/ITunesSearchAPI.kt
@@ -1,0 +1,13 @@
+package com.practicum.playlistmaker
+
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+
+interface ITunesSearchAPI {
+
+    @GET("/search?entity=song")
+    fun search(@Query("term") text: String): Call<TracksResponse>
+
+}
