+++ app/src/main/java/com/practicum/playlistmaker/MainActivity.kt
@@ -1,14 +1,10 @@
-package com.practicum.playlistmaker.presentation.main
+package com.practicum.playlistmaker
 
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
-import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.presentation.search.SearchActivity
-import com.practicum.playlistmaker.presentation.settings.SettingsActivity
-import com.practicum.playlistmaker.presentation.library.MediaLibraryActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
