+++ app/src/main/java/com/practicum/playlistmaker/MediaLibraryActivity.kt
@@ -1,10 +1,9 @@
-package com.practicum.playlistmaker.presentation.library
+package com.practicum.playlistmaker
 
 import android.os.Bundle
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import com.bumptech.glide.Glide
-import com.practicum.playlistmaker.R
 
 class MediaLibraryActivity : AppCompatActivity() {
