+++ app/src/main/java/com/practicum/playlistmaker/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.presentation.settings
+package com.practicum.playlistmaker
 
 import android.content.Intent
 import android.net.Uri
@@ -7,13 +7,8 @@ import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.appbar.MaterialToolbar
 import com.google.android.material.switchmaterial.SwitchMaterial
 import com.google.android.material.textview.MaterialTextView
-import com.practicum.playlistmaker.Creator.Creator
-import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.domain.api.SettingsThemeInteractor
 
 class SettingsActivity : AppCompatActivity() {
-
-    private lateinit var settingsInteractor: SettingsThemeInteractor
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -30,13 +25,10 @@ class SettingsActivity : AppCompatActivity() {
             finish()
         }
 
-        settingsInteractor = Creator.provideSettingsThemeInteractor()
+        val app = applicationContext as App
+        buttonSwitch.isChecked = app.darkTheme
 
-        buttonSwitch.isChecked = settingsInteractor.isDarkThemeEnabled()
-
-        buttonSwitch.setOnCheckedChangeListener { switcher, checked ->
-            settingsInteractor.setTheme(checked)
-        }
+        buttonSwitch.setOnCheckedChangeListener { switcher, checked ->  app.switchTheme(checked)}
 
         buttonShareApp.setOnClickListener {
             val linkApp = getString(R.string.linkToShareApp)
