+++ app/src/main/java/com/practicum/playlistmaker/Track.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.domain.model
+package com.practicum.playlistmaker
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
@@ -6,7 +6,7 @@ import java.text.SimpleDateFormat
 import java.util.Locale
 
 @Parcelize
-data class Tracks(
+data class Track(
     val trackName: String,
     val artistName: String,
     val trackTimeMillis: Long,
@@ -16,12 +16,10 @@ data class Tracks(
     val releaseDate: String,
     val primaryGenreName: String,
     val country: String,
-    val previewUrl: String,
-) : Parcelable {
+): Parcelable {
     val formattedTrackTime: String
         get() = SimpleDateFormat("mm:ss", Locale.getDefault()).format(trackTimeMillis)
 
     val getCoverArtwork: String
         get() = artworkUrl100.replaceAfterLast("/", "512x512bb.jpg")
-
 }
