+++ app/src/main/java/com/practicum/playlistmaker/TrackAdapter.kt
@@ -1,14 +1,10 @@
-package com.practicum.playlistmaker.presentation.search
+package com.practicum.playlistmaker
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import com.practicum.playlistmaker.domain.model.Tracks
 import kotlin.collections.ArrayList
 
-class TrackAdapter(
-    private var tracks: ArrayList<Tracks>,
-    private val clickListener: ClickListener
-) : RecyclerView.Adapter<TrackViewHolder>() {
+class TrackAdapter(private var tracks: ArrayList<Track>, private val clickListener: ClickListener): RecyclerView.Adapter<TrackViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): TrackViewHolder {
         return TrackViewHolder(parent)
@@ -21,18 +17,18 @@ class TrackAdapter(
     override fun onBindViewHolder(holder: TrackViewHolder, position: Int) {
         holder.bind(tracks[position])
 
-        holder.itemView.setOnClickListener {
+        holder.itemView.setOnClickListener{
             val track = tracks[position]
             clickListener.onItemClick(track)
         }
     }
 
-    fun updateData(track: ArrayList<Tracks>) {
+    fun updateData(track: ArrayList<Track>){
         this.tracks = track
         notifyDataSetChanged()
     }
 
-    fun interface ClickListener {
-        fun onItemClick(track: Tracks)
+    fun interface ClickListener{
+        fun onItemClick(track: Track)
     }
 }
