+++ app/src/main/java/com/practicum/playlistmaker/TrackViewHolder.kt
@@ -1,4 +1,4 @@
-package com.practicum.playlistmaker.presentation.search
+package com.practicum.playlistmaker
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,21 +7,16 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.domain.model.Tracks
-import com.practicum.playlistmaker.Utils.dpToPx
 
-class TrackViewHolder(parent: ViewGroup) :
-    RecyclerView.ViewHolder(
-        LayoutInflater.from(parent.context).inflate(R.layout.search_track_view, parent, false)
-    ) {
+class TrackViewHolder(parent: ViewGroup):
+    RecyclerView.ViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.search_track_view, parent, false)) {
 
     private val ivCoverTrack = itemView.findViewById<ImageView>(R.id.ivCoverTrack)
     private val tvTrackName = itemView.findViewById<TextView>(R.id.tvTrackName)
     private val tvArtistName = itemView.findViewById<TextView>(R.id.tvArtistName)
     private val tvTrackTime = itemView.findViewById<TextView>(R.id.tvTrackDuration)
 
-    fun bind(model: Tracks) {
+    fun bind (model: Track){
         Glide.with(itemView)
             .load(model.artworkUrl100)
             .placeholder(R.drawable.placeholder)
