+++ app/src/main/java/com/practicum/playlistmaker/TracksResponse.kt
@@ -1,0 +1,4 @@
+package com.practicum.playlistmaker
+
+data class TracksResponse (val resultCount: Int,
+                           val results: List<Track>)
