+++ app/src/main/java/com/practicum/playlistmaker/Utils.kt
@@ -1,0 +1,12 @@
+package com.practicum.playlistmaker
+
+import android.content.Context
+import android.util.TypedValue
+
+fun dpToPx(dp: Float, context: Context): Int {
+    return TypedValue.applyDimension(
+        TypedValue.COMPLEX_UNIT_DIP,
+        dp,
+        context.resources.displayMetrics
+    ).toInt()
+}
