--- app/src/main/java/com/practicum/playlistmaker/Utils/Utils.kt
@@ -1,24 +1,0 @@
-package com.practicum.playlistmaker.Utils
-
-import android.content.Context
-import android.util.TypedValue
-import com.practicum.playlistmaker.data.dto.TracksDto
-import com.practicum.playlistmaker.domain.model.Tracks
-
-fun dpToPx(dp: Float, context: Context): Int {
-    return TypedValue.applyDimension(
-        TypedValue.COMPLEX_UNIT_DIP,
-        dp,
-        context.resources.displayMetrics
-    ).toInt()
-}
-
-fun Tracks.toDto(): TracksDto = TracksDto(
-    trackName, artistName, trackTimeMillis, artworkUrl100, trackId, collectionName, releaseDate,
-    primaryGenreName, country, previewUrl
-)
-
-fun TracksDto.toDomain(): Tracks = Tracks(
-    trackName, artistName, trackTimeMillis, artworkUrl100, trackId, collectionName, releaseDate,
-    primaryGenreName, country, previewUrl
-)
