--- app/src/main/java/com/practicum/playlistmaker/data/dto/TracksDto.kt
@@ -1,14 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-data class TracksDto(
-    val trackName: String,
-    val artistName: String,
-    val trackTimeMillis: Long,
-    val artworkUrl100: String,
-    val trackId: Int,
-    val collectionName: String?,
-    val releaseDate: String,
-    val primaryGenreName: String,
-    val country: String,
-    val previewUrl: String,
-)
