--- app/src/main/java/com/practicum/playlistmaker/data/dto/TracksSearchResponse.kt
@@ -1,6 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-data class TracksSearchResponse(
-    val resultCount: Int,
-    val results: List<TracksDto>
-) : Response()
