--- app/src/main/java/com/practicum/playlistmaker/data/network/ITunesSearchAPI.kt
@@ -1,13 +1,0 @@
-package com.practicum.playlistmaker.data.network
-
-import com.practicum.playlistmaker.data.dto.TracksSearchResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ITunesSearchAPI {
-
-    @GET("/search?entity=song")
-    fun search(@Query("term") text: String): Call<TracksSearchResponse>
-
-}
