--- app/src/main/java/com/practicum/playlistmaker/data/network/RetrofitNetworkClientImpl.kt
@@ -1,36 +1,0 @@
-package com.practicum.playlistmaker.data.network
-
-import com.practicum.playlistmaker.data.NetworkClient
-import com.practicum.playlistmaker.data.dto.Response
-import com.practicum.playlistmaker.data.dto.TracksSearchRequest
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import java.io.IOException
-
-class RetrofitNetworkClientImpl : NetworkClient {
-
-    private val iTunesBaseUrl = "https://itunes.apple.com"
-
-    private val retrofit = Retrofit.Builder()
-        .baseUrl(iTunesBaseUrl)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    private val iTunesService = retrofit.create(ITunesSearchAPI::class.java)
-
-    override fun doRequest(dto: Any): Response {
-        return try {
-            if (dto is TracksSearchRequest) {
-                val resp = iTunesService.search(dto.expression).execute()
-                val body = resp.body() ?: Response()
-                body.apply { resultCode = resp.code() }
-            } else {
-                Response().apply { resultCode = 400 }
-            }
-        } catch (e: IOException) {
-            Response().apply { resultCode = -1 }
-        } catch (e: Exception) {
-            Response().apply { resultCode = 500 }
-        }
-    }
-}
