--- app/src/main/java/com/practicum/playlistmaker/domain/api/AudioPlayerRepository.kt
@@ -1,20 +1,0 @@
-package com.practicum.playlistmaker.domain.api
-
-interface AudioPlayerRepository {
-    fun preparePlayer(
-        url: String,
-        onPrepared: () -> Unit,
-        onCompletion: () -> Unit
-    )
-
-    fun startPlayer(onStarted: () -> Unit)
-    fun pausePlayer(onPaused: () -> Unit)
-    fun playbackControl(
-        onStarted: () -> Unit = {},
-        onPaused: () -> Unit = {}
-    )
-
-    fun releasePlayer()
-    fun getCurrentPosition(): Int
-    fun isPlaying(): Boolean
-}
