--- app/src/main/java/com/practicum/playlistmaker/domain/api/SettingsThemeInteractor.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.domain.api
-
-interface SettingsThemeInteractor {
-    fun isDarkThemeEnabled(): Boolean
-    fun setTheme(darkTheme: Boolean)
-    fun isSystemThemeSet(): Boolean
-}
