--- app/src/main/java/com/practicum/playlistmaker/domain/api/SettingsThemeRepository.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.domain.api
-
-interface SettingsThemeRepository {
-    fun getTheme(): Boolean
-    fun saveTheme(darkTheme: Boolean)
-    fun isSystemThemeSet(): Boolean
-}
