--- app/src/main/java/com/practicum/playlistmaker/domain/model/PlayerState.kt
@@ -1,8 +1,0 @@
-package com.practicum.playlistmaker.domain.model
-
-enum class PlayerState {
-    DEFAULT,
-    PREPARED,
-    PLAYING,
-    PAUSED
-}
