--- app/src/main/java/com/practicum/playlistmaker/domain/model/ResultSearchTracks.kt
@@ -1,6 +1,0 @@
-package com.practicum.playlistmaker.domain.model
-
-sealed class ResultSearchTracks {
-    data class Success(val tracks: List<Tracks>) : ResultSearchTracks()
-    data class Error(val  message: String) : ResultSearchTracks()
-}
