--- app/src/main/res/drawable-night/pause_button.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="100dp"
-    android:height="100dp"
-    android:viewportWidth="100"
-    android:viewportHeight="100">
-  <path
-      android:pathData="M50,8.33C27,8.33 8.33,27 8.33,50C8.33,73 27,91.67 50,91.67C73,91.67 91.67,73 91.67,50C91.67,27 73,8.33 50,8.33ZM45.83,66.67H37.5V33.33H45.83V66.67ZM62.5,66.67H54.17V33.33H62.5V66.67Z"
-      android:fillColor="#ffffff"/>
-</vector>
