+++ app/src/main/res/layout/activity_media_library.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.library.MediaLibraryActivity"
+    tools:context=".MediaLibraryActivity"
     android:orientation="vertical"
     android:gravity="center_horizontal">
