+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.search.SearchActivity">
+    tools:context=".SearchActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/arrowBack"
@@ -86,15 +86,6 @@
         android:layout_height="wrap_content"
         android:orientation="vertical">
 
-        <ProgressBar
-            android:id="@+id/progressBar"
-            android:layout_width="@dimen/size_for_progress_bar"
-            android:layout_height="@dimen/size_for_progress_bar"
-            android:layout_gravity="center"
-            android:layout_marginTop="140dp"
-            android:indeterminateTint="@color/dark_blue"
-            android:visibility="gone" />
-
         <TextView
             android:id="@+id/tvSearchHistory"
             style="@style/SearchResultText"
@@ -113,7 +104,6 @@
             android:layout_marginTop="16dp"
             android:layout_weight="1"
             android:orientation="vertical"
-            android:visibility="visible"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             tools:itemCount="5"
             tools:listitem="@layout/search_track_view" />
