+++ app/src/main/res/values/dimens.xml
@@ -7,7 +7,4 @@
     <dimen name="button_radius">16dp</dimen>
     <dimen name="settings_text_size">16sp</dimen>
     <dimen name="settings_lists_padding">21dp</dimen>
-    <dimen name="size_for_progress_bar">44dp</dimen>
-    <dimen name="player_like_and_add_icon">51dp</dimen>
-    <dimen name="play_or_pause_btn">100dp</dimen>
 </resources>
