+++ app/src/main/res/values/strings.xml
@@ -29,7 +29,5 @@
     <string name="year">Год</string>
     <string name="album">Альбом</string>
     <string name="duration">Длительность</string>
-    
-    <string name="track_time_start">00:00</string>
 
 </resources>
